fix(ast/32-bit): stop gen_cast's 64->32 split desyncing the recorder; wire i386 optfire
Wiring the i386 optfire cells left 6 red, which split into two independent
problems that the previous triage had merged into one:

1. narrow_fix/narrow_class0..3 were written with `long` -- 64-bit on x86_64
   but 32-bit on i386, so there was no 64->32 narrowing to find there at all.
   Those functions were fully faithful on i386; the pass simply had nothing to
   do. Rewritten in `long long`, which is 64-bit on every target mcc supports.

2. A real compiler defect, which vlat (already `long long`) was exposing.
   gen_cast splits a 64-bit value into 32-bit halves via lexpand/lbuild on
   32-bit targets, and lexpand's vdup ran through ast_hook_vpush, which saw a
   bare VT_LLONG register value it cannot model and set desync -- so EVERY
   64->32 narrowing desynced the whole function and the AST optimizer silently
   abandoned it. Same class as the gaddrof materialisation bug, same fix:
   bracket the split with ast_hook_synth_begin/_end. Safe because the cast is
   already modelled -- ast_hook_convert records an AST_Convert at gen_cast
   entry, so replay re-emits it; the lexpand is emission detail.

On i386 vlat goes 3 desync + 1 faithful -> 4 faithful, and narrow 0 -> 3,
matching x86_64 exactly. optfire-i386 is now 45/45 (was 39/45).

Validation, since this changes real 32-bit codegen: x86_64 unaffected (0 of
750 objects differ, -O0/-O2/-O3, pre-fix vs post-fix compiler); i386 3/244 and
arm 2/239 objects change; native i386 differential vs gcc -m32 over 227 exec
programs at -O0/-O2 with 0 regressions; qemu-arm 10/10; qemu-i386 10/10; host
ctest 7229/7229; cross ctest 7343/7343; self-host fixpoint byte-identical.

+++ tests/optfire/src/narrow_class0.c
@@ -1,17 +1,17 @@
 extern int printf(const char *, ...);
 
-static int chained(int a, long c, long d)
+static int chained(int a, long long c, long long d)
 {
-	return (int)((int)((long)a * c) + d);
+	return (int)((int)((long long)a * c) + d);
 }
 
-static int chained2(int a, long c, long d)
+static int chained2(int a, long long c, long long d)
 {
-	return (int)((int)((long)a + c) - d);
+	return (int)((int)((long long)a + c) - d);
 }
 
 int main(void)
 {
-	printf("narrow_class0=%d\n", chained(7, 5L, 11L) + chained2(3, 4L, 2L));
+	printf("narrow_class0=%d\n", chained(7, 5LL, 11LL) + chained2(3, 4LL, 2LL));
 	return 0;
 }
