+++ tests/optfire/src/narrow_fix.c
@@ -2,12 +2,12 @@ extern int printf(const char *, ...);
 
 static int triple(int a, int b, int c)
 {
-	return (int)((long)a + (long)b + (long)c);
+	return (int)((long long)a + (long long)b + (long long)c);
 }
 
 static int quad(int a, int b, int c, int d)
 {
-	return (int)((long)a + (long)b + (long)c + (long)d);
+	return (int)((long long)a + (long long)b + (long long)c + (long long)d);
 }
 
 int main(void)
